+++ wrappers/python/src/cxx/CMakeLists.txt
@@ -46,6 +46,7 @@ set(SOURCE_FILES
         common/SpdlogPythonSink.h
         common/Utils.cpp
         api/ExportObservables.h
+        api/PyTopology.h
         api/ExportLoopApi.cpp
         api/ExportTopologies.cpp
         api/ExportKernelContext.cpp
