+++ wrappers/python/src/cxx/api/ApiModule.cpp
@@ -42,6 +42,7 @@
 #include <readdy/model/actions/UserDefinedAction.h>
 #include "ExportObservables.h"
 #include "../common/ReadableParticle.h"
+#include "PyTopology.h"
 
 namespace py = pybind11;
 
@@ -104,7 +105,8 @@ void exportApi(py::module &api) {
             .def("get_particles_for_topology", &sim::getParticlesForTopology, "topology"_a)
             .def("add_topology", [](sim &self, const std::string &name,
                                     const std::vector<readdy::model::Particle> &particles) {
-                return self.addTopology(name, particles);
+                auto* top = self.addTopology(name, particles);
+                return PyTopology(top);
             }, rvp::reference_internal, "type"_a, "particles"_a)
             .def("add_topology", [](sim &self, const std::string &name, const std::vector<std::string> &types,
                                     const py::array_t<readdy::scalar> &positions) {
@@ -121,9 +123,15 @@ void exportApi(py::module &api) {
                                                                                        positions.at(i, 1),
                                                                                        positions.at(i, 2))));
                 }
-                return self.addTopology(name, particles);
+                return PyTopology(self.addTopology(name, particles));
             }, rvp::reference_internal)
-            .def_property_readonly("current_topologies", &sim::currentTopologies)
+            .def_property_readonly("current_topologies", [](sim &self) {
+                auto curr = self.currentTopologies();
+                std::vector<PyTopology> topologies;
+                topologies.reserve(curr.size());
+                std::transform(curr.begin(), curr.end(), std::back_inserter(topologies), [](auto* ptr) { return PyTopology(ptr); });
+                return topologies;
+            })
             .def_property_readonly("current_particles", [](const sim &self) {
                 std::vector<rpy::ReadableParticle> particles;
                 auto currentParticles = self.currentParticles();
@@ -142,7 +150,6 @@ void exportApi(py::module &api) {
                 self.run(steps, timeStep);
             }, "n_steps"_a, "time_step"_a);
     exportObservables(api, simulation);
-
 
     struct nodelete {
         void operator()(kp* ptr) const {}
